Remove serial traits (blocking and async).
