+++ embedded-hal/src/lib.rs
@@ -81,7 +81,6 @@ pub mod delay;
 pub mod digital;
 pub mod i2c;
 pub mod pwm;
-pub mod serial;
 pub mod spi;
 
 mod private {
